+++ packages/ui/src/index.ts
@@ -1,1 +1,2 @@
 export * from './button/src'
+export * from './checkboxInput/src'