+++ packages/ui/CHANGELOG.md
@@ -10,7 +10,7 @@
 
 ### Patch Changes
 
-- 9a2b4bd: scope namescape for testing library
+- 9a2b4bd: scope namespace for testing library
 
 ## 1.0.0